requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
colorama==0.4.6
distro==1.9.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.37.2
pydantic==2.8.2
pydantic_core==2.20.1
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dotenv==1.0.1
setuptools==71.1.0
shiboken6==6.7.2
sniffio==1.3.1
SQLAlchemy==2.0.31
SQLite3-0611==0.0.1
tqdm==4.66.5
typing_extensions==4.12.2
